Moved parsing functions to a separate file
