+++ src/criminal-code-offence-parser/constants.py
@@ -57,3 +57,8 @@ VIOLENCE_USED_THREATENED_ATTEMPTED_OFFENCES = [
     "cc46(1)", "cc46(2)", "cc65(1)", "cc65(2)", "cc74", "cc75", "cc76", "cc78.1(4)",
     "cc87", "cc88", 
 ]
+
+SECTION_469_OFFENCES = [
+    "cc46(1)", "cc46(2)", "cc51", "cc53", "cc61", "cc74", "cc75", "cc231(2)", "cc231(7)", 
+    "cc119",
+]
