+++ src/criminal-code-offence-parser/data/cc-offences-2024-09-16.csv
@@ -240,7 +240,7 @@ cc247(3),traps likely to cause bodily harm — offence-related place,,10y,,sc
 cc247(4),traps likely to cause bodily harm — offence-related place — causing bodily harm ,,14y,,
 cc247(5),traps likely to cause bodily harm — death,,999y,,
 cc248,interfering with transportation facility,,999y,,
-cc262,impeding attempt to save 999y,,10y,,sc
+cc262,impeding attempt to save life,,10y,,sc
 cc263(3)(c),offences — duty to safeguard opening in ice or excavation on land,,,,sc
 cc264,criminal harassment,,10y,,sc
 cc264.1(a),uttering threats — death or bodily harm,,5y,,sc
@@ -522,7 +522,7 @@ cc459,conveying instruments for coining out of mint,,14y,,
 cc460,advertising and dealing in counterfeit money,,5y,,sc
 cc462.31(1),laundering proceeds of crime,,10y,,sc
 cc462.31(2.1),laundering proceeds of crime — criminal organization,,14y,,
-cc463(a),attempts accessories — punishable by 999y,,14y,,
+cc463(a),attempts accessories — punishable by life,,14y,,
 cc463(b),attempts accessories — punishable by 14y or less,,7y,,
 cc463(c),attemps accessories — sc conviction,,,,sc
 cc463(d),attempts accessories — hybrid,,7y,,sc
